ci: improve error handling
