+++ scripts/release/index.js
@@ -1,16 +1,6 @@
 const yargs = require("yargs");
-const { LatestRelease } = require("./LatestRelease");
-const { BetaRelease } = require("./BetaRelease");
-const { UnstableRelease } = require("./UnstableRelease");
-const { VerdaccioRelease } = require("./VerdaccioRelease");
 const { ConsoleLogger } = require("../ConsoleLogger");
-
-const releaseTypes = {
-    latest: LatestRelease,
-    beta: BetaRelease,
-    unstable: UnstableRelease,
-    verdaccio: VerdaccioRelease
-};
+const { getReleaseType } = require("./releaseTypes");
 
 (async () => {
     const { type, tag } = yargs.argv;
@@ -18,12 +8,7 @@ const releaseTypes = {
         throw Error(`Missing required "--type" option.`);
     }
 
-    const Release = releaseTypes[type];
-
-    if (!Release) {
-        const possibleTypes = Object.keys(releaseTypes).join(", ");
-        throw Error(`Unrecognized type "${type}". Specify one of: ${possibleTypes}`);
-    }
+    const Release = getReleaseType(type);
 
     const logger = new ConsoleLogger();
     const release = new Release(logger);
