+++ scripts/release/triggerRelease.js
@@ -1,30 +1,34 @@
 #!/usr/bin/env node
+const yargs = require("yargs");
 const { Octokit } = require("@octokit/rest");
 const { ConsoleLogger } = require("../ConsoleLogger");
-const yargs = require("yargs");
+const { checkReleaseType } = require("./releaseTypes");
 
 /**
- * A simple script that will trigger a "release" GitHub workflow.
+ * A simple script that will trigger the "release" GitHub workflow.
  */
 
 (async () => {
-    const logger = new ConsoleLogger();
     const { branch, type, token, tag } = yargs.argv;
 
-    if (!branch) {
-        throw Error(`"--branch" argument missing. Specify a branch you want to release.`);
-    }
-
-    if (!type) {
-        throw Error(`"--type" argument missing. Specify one of: latest, beta, unstable.`);
-    }
-
     try {
+        if (!branch) {
+            throw Error(`"--branch" argument missing. Specify a branch you want to release.`);
+        }
+
+        if (!type) {
+            throw Error(`"--type" argument missing. Specify one of: latest, beta, unstable.`);
+        }
+
+        checkReleaseType(type);
+
         if (!token) {
-            throw new Error(
+            throw Error(
                 `"--token" argument missing. Make sure it contains a valid GitHub access token.`
             );
         }
+
+        const logger = new ConsoleLogger();
 
         logger.info("Branch: %s", branch);
         logger.info("Type: %s", type);
@@ -48,7 +52,6 @@ const yargs = require("yargs");
         logger.success("GitHub workflow triggered successfully!");
         logger.info("See action details: %s", "https://github.com/webiny/webiny-js/actions");
     } catch (e) {
-        console.log(e);
-        logger.error("Something went wrong: %s", e.message);
+        logger.error(e.message);
     }
 })();
